Changed snowmanBlocks to be pointers to simplify the implementation
